docker_practice_app/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
redis==5.0.6
